Moved the SolverStats class to the solver file Renamed the class Stats, as it seemed my in keeping with the other classes in the file Added a `clear` method
